requirements.txt
openai==0.28.0
faiss-cpu==1.7.4
pandas==2.1.0
PyMuPDF==1.24.1
python-dotenv==1.0.0
numpy==1.24.3
